chore: features 폴더 auth 생성
