+++ src/features/auth/model/type.ts
@@ -1,0 +1,12 @@
+export interface LoginFormTypes {
+  email: string;
+  password: string;
+}
+
+export interface RegisterFormTypes {
+  nickname: string;
+  companyName: string;
+  email: string;
+  password: string;
+  passwordCheck: string;
+}
