--- src/features/auth/type.ts
@@ -1,12 +1,0 @@
-export interface RegisterFormTypes {
-  nickname: string;
-  companyName: string;
-  email: string;
-  password: string;
-  passwordCheck: string;
-}
-
-export interface LoginFormTypes {
-  email: string;
-  password: string;
-}
