+++ src/shared/components/input/inputWithError.tsx
@@ -1,0 +1,6 @@
+import withError from '@/shared/lib/HOC/withError';
+import Input from '.';
+
+const InputWithMessage = withError(Input);
+
+export default InputWithMessage;
