+++ src/shared/components/input/inputWithLabel.tsx
@@ -1,0 +1,6 @@
+import withLabel from '@/shared/lib/HOC/withLabel';
+import InputWithMessage from './inputWithError';
+
+const InputWithLabel = withLabel(InputWithMessage);
+
+export default InputWithLabel;
