+++ src/shared/lib/HOC/withError.tsx
